app/requirements.txt
fastapi
uvicorn
tensorflow
scikit-learn
pymysql
python-dotenv
python-jose